requirements-dev.txt
django-debug-toolbar
mypy
setuptools
wheel
